feat: HomeTodayTopicScreen.kt cursorId 기반 페이징 구현

+++ app/src/main/java/com/lyfe/android/core/domain/repository/BoardRepository.kt
@@ -3,7 +3,6 @@ package com.lyfe.android.core.domain.repository
 import com.lyfe.android.core.data.model.GetBoardDetailResponse
 import com.lyfe.android.core.data.network.model.Result
 import com.lyfe.android.core.model.Feed
-import com.lyfe.android.core.model.Page
 import kotlinx.coroutines.flow.Flow
 
 interface BoardRepository {
@@ -11,12 +10,12 @@ interface BoardRepository {
 	suspend fun getBoardDetail(boardId: Long): Result<GetBoardDetailResponse>
 
 	fun getLatestBoards(
-		cursorId: Int,
+		cursorId: Long,
 		boardType: String
 	): Flow<List<Feed>>
 
 	fun getPopularBoards(
-		cursorId: Int = 0,
+		cursorId: Long = 0,
 		boardType: String?,
 		popularType: String
 	): Flow<List<Feed>>
