+++ app/src/main/java/com/lyfe/android/core/domain/usecase/GetPopularBoardsUseCase.kt
@@ -8,7 +8,7 @@ class GetPopularBoardsUseCase @Inject constructor(
 ) {
 
 	operator fun invoke(
-		cursorId: Int = 0,
+		cursorId: Long = 0,
 		boardType: String?,
 		popularType: String
 	) = boardRepository.getPopularBoards(cursorId, boardType, popularType)
